Assessment UX — gradient sliders, live score ring, timeline history
Replaces the basic 1-10 integer number inputs with a modern slider
interface: red→amber→green gradient track, 0.5-point steps, live SVG
overall-score ring, and a full-width submit button disabled until every
category is scored. History is now a timeline with colored left borders
(improved/declined/first), trend arrows, mini per-category bars, and an
expandable color-coded chip grid. Applied to both the coach's
AssessmentPage and the athlete's My Stats page.

Backend: PlayerStatScore.Score is now decimal(3,1) instead of int, since
the UI can now submit half-point scores — int caused a 400 on submit.
Also added a global UTC DateTime value converter on ApplicationDbContext:
Postgres' timestamp with time zone rejects Kind=Unspecified values (e.g.
a plain <input type="date"> bound into a DTO), which surfaced as a 500 on
assessment submission after the earlier SQLite→Postgres migration.

+++ Dtos/AssessmentDtos.cs
@@ -43,12 +43,12 @@ public class PlayerStatScoreDto
     public int PlayerAssessmentId { get; set; }
     public int SportStatCategoryId { get; set; }
     public string StatCategoryName { get; set; } = "";
-    public int Score { get; set; }
+    public decimal Score { get; set; }
 }
 
 public class CreatePlayerStatScoreDto
 {
     public int PlayerAssessmentId { get; set; }
     public int SportStatCategoryId { get; set; }
-    public int Score { get; set; }
+    public decimal Score { get; set; }
 }
